Peer-to-peer handshake (#47)
* Paid tech debt

* Intoduced Handshake

* Added Serialize/Deserialize

* Added messaging

* Dropped handshake strategy

* Made it slimmer for messages

* Moved version initilization

* Added handshake into tests

+++ src/NeoSharp.Core/Network/ConfigurationExtensions.cs
@@ -18,9 +18,9 @@ namespace NeoSharp.Core.Network
             networkConfig.PeerEndPoints = ParsePeerEndPoints(config);
         }
 
-        private static int ParsePort(IConfiguration config)
+        private static ushort ParsePort(IConfiguration config)
         {
-            var port = config.GetSection("port")?.Get<int>();
+            var port = config.GetSection("port")?.Get<ushort>();
             return port ?? 0;
         }
